requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
authlib==1.3.0
websockets==12.0
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
python-dotenv==1.0.1
alembic==1.13.1
email-validator==2.1.0
google-auth==2.28.1
